+++ src/main/java/org/example/mybatis/controller/VideoController.java
@@ -28,21 +28,21 @@ public class VideoController {
 
 
     @DeleteMapping("/async/{videoId}")
-    public CompletableFuture<ResponseEntity<String>> deleteVideoAsync(@PathVariable("videoId") int videoId) {
+    public CompletableFuture<ResponseEntity<String>> deleteVideoAsync(@PathVariable("videoId") Long videoId) {
         return asyncVideoService.deleteVideoProcedure(videoId)
                 .thenApply(aVoid -> new ResponseEntity<>("Video deleted successfully", HttpStatus.NO_CONTENT));
     }
 
     @PutMapping("/async/{videoId}")
     public CompletableFuture<ResponseEntity<String>> updateVideoInfoAsync(
-            @PathVariable("videoId") int videoId,
+            @PathVariable("videoId") Long videoId,
             @RequestBody Video video) {
         return asyncVideoService.updateVideoInfoProcedure(videoId, video.getTitle(), video.getDescription(), video.getVideoPath())
                 .thenApply(aVoid -> new ResponseEntity<>("Video updated successfully", HttpStatus.OK));
     }
 
     @GetMapping("/async/{userId}")
-    public CompletableFuture<ResponseEntity<List<Video>>> getUserVideosAsync(@PathVariable("userId") int userId) {
+    public CompletableFuture<ResponseEntity<List<Video>>> getUserVideosAsync(@PathVariable("userId") Long userId) {
         return asyncVideoService.getUserVideosProcedure(userId)
                 .thenApply(videos -> {
                     if (videos != null) {
@@ -63,7 +63,7 @@ public class VideoController {
 
     // 通过 ID 查询单个用户
     @GetMapping("/{videoId}")
-    public ResponseEntity<Video> getVideoByVideoId(@PathVariable("videoId") int videoId) {
+    public ResponseEntity<Video> getVideoByVideoId(@PathVariable("videoId") Long videoId) {
         Video video = videoService.getVideoByVideoId(videoId);
         if (video != null) {
             return new ResponseEntity<>(video, HttpStatus.OK);
@@ -81,7 +81,7 @@ public class VideoController {
 
     // 更新用户信息
     @PutMapping("/{videoId}")
-    public ResponseEntity<Video> updateVideo(@PathVariable("videoId") int videoId, @RequestBody Video video) {
+    public ResponseEntity<Video> updateVideo(@PathVariable("videoId") Long videoId, @RequestBody Video video) {
         video.setVideoID(videoId);
         Video updatedVideo = videoService.updateVideo(video);
         if (updatedVideo != null) {
@@ -93,7 +93,7 @@ public class VideoController {
 
     // 删除用户
     @DeleteMapping("/{videoId}")
-    public ResponseEntity<Void> deleteVideo(@PathVariable("videoId") int videoId) {
+    public ResponseEntity<Void> deleteVideo(@PathVariable("videoId") Long videoId) {
         boolean isDeleted = videoService.deleteVideo(videoId);
         if (isDeleted) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
