+++ src/main/java/org/example/mybatis/service/VideoService.java
@@ -8,11 +8,11 @@ import java.util.List;
 public interface VideoService {
     List<Video> listAllVideos();
 
-    Video getVideoByVideoId(int videoId);
+    Video getVideoByVideoId(Long videoId);
 
     void insertVideo(Video video);
 
     Video updateVideo(Video video);
 
-    boolean deleteVideo(int videoID);
+    boolean deleteVideo(Long videoID);
 }
